move prometheus package to shared directory
